# assets/swagger-v2-consumers.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Civitai Orchestration Consumer API",
    "description": "The Civitai Orchestration API is a RESTful API that allows you to interact with the Civitai Orchestration system. The API is designed to be simple and easy to use, and is built on top of the OpenAPI specification.",
    "version": "v2",
    "x-logo": {
      "url": "https://wiki.civitai.com/images/1/17/Logo_%28Light%29.png",
      "backgroundColor": "#FFFFFF"
    }
  },
  "paths": {
...
    "/v2/resources/{air}": {
      "get": {
        "tags": [
          "Resources"
        ],
        "summary": "Get an existing resource.",
        "operationId": "GetResource",
        "parameters": [
          {
            "name": "air",
            "in": "path",
            "description": "A unique ID for the resource being requested. See https://developer.civitai.com/docs/getting-started/ai-resource-identifier for more info on AIRs.",
            "required": true,
            "schema": {
              "pattern": "^(?:urn:)?(?:air:)?(?:(?<ecosystem>[a-zA-Z0-9_\\-\\/]+):)?(?:(?<type>[a-zA-Z0-9_\\-\\/]+):)?(?<source>[a-zA-Z0-9_\\-\\/]+):(?<id>[a-zA-Z0-9_\\-\\/\\.]+)(?:@(?<version>[a-zA-Z0-9_\\-\\/.=,]+))?(?:\\.(?<format>[a-zA-Z0-9_\\-]+))?$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Resources"
        ],
        "summary": "Invalidates the cache of a specific resource.",
        "operationId": "InvalidateResource",
        "parameters": [
          {
            "name": "air",
            "in": "path",
            "description": "A unique ID for the resource being requested. See https://developer.civitai.com/docs/getting-started/ai-resource-identifier for more info on AIRs.",
            "required": true,
            "schema": {
              "pattern": "^(?:urn:)?(?:air:)?(?:(?<ecosystem>[a-zA-Z0-9_\\-\\/]+):)?(?:(?<type>[a-zA-Z0-9_\\-\\/]+):)?(?<source>[a-zA-Z0-9_\\-\\/]+):(?<id>[a-zA-Z0-9_\\-\\/\\.]+)(?:@(?<version>[a-zA-Z0-9_\\-\\/.=,]+))?(?:\\.(?<format>[a-zA-Z0-9_\\-]+))?$",
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "One or more userIds to invalidate early access for",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "etag",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FileFormat": {
        "enum": [
          "unknown",
          "safeTensor",
          "pickleTensor",
          "diffusers",
          "coreML",
          "onnx"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ResourceInfo": {
        "required": [
          "air",
          "downloadUrls",
          "hashes",
          "size"
        ],
        "type": "object",
        "properties": {
          "air": {
            "pattern": "^(?:urn:)?(?:air:)?(?:(?<ecosystem>[a-zA-Z0-9_\\-\\/]+):)?(?:(?<type>[a-zA-Z0-9_\\-\\/]+):)?(?<source>[a-zA-Z0-9_\\-\\/]+):(?<id>[a-zA-Z0-9_\\-\\/\\.]+)(?:@(?<version>[a-zA-Z0-9_\\-\\/.=,]+))?(?:\\.(?<format>[a-zA-Z0-9_\\-]+))?$",
            "type": "string",
            "description": "An AIR ID for the resource."
          },
          "size": {
            "type": "integer",
            "description": "The resource size in bytes.",
            "format": "int64"
          },
          "hashes": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "A collection of hashes."
          },
          "downloadUrls": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "description": "An array of download urls."
          },
          "resourceName": {
            "type": "string",
            "description": "The name of the resource.",
            "nullable": true
          },
          "versionName": {
            "type": "string",
            "description": "The name of the version.",
            "nullable": true
          },
          "invalidateAt": {
            "type": "string",
            "description": "The date time to invalidate at.",
            "format": "date-time",
            "nullable": true
          },
          "earlyAccessEndsAt": {
            "type": "string",
            "description": "A DateTime representing when early access for the resource ends.",
            "format": "date-time",
            "nullable": true
          },
          "checkPermission": {
            "type": "boolean",
            "description": "A bool indicating if permission is required to use this resource."
          },
          "canGenerate": {
            "type": "boolean",
            "description": "A bool indicating if generation is enabled for this resource."
          },
          "freeTrialLimit": {
            "type": "integer",
            "description": "An optional limit on the number of uses for this resource per user that has early acccess.",
            "format": "int32",
            "nullable": true
          },
          "requiresAuthorization": {
            "type": "boolean",
            "description": "Wether this resource requires authorization.",
            "nullable": true
          },
          "fileFormat": {
            "$ref": "#/components/schemas/FileFormat"
          },
          "hasMatureContentRestriction": {
            "type": "boolean",
            "description": "A boolean indicating whether this resource restricts mature content generation.\r\nIf resources with this restriction are used in generation, then generations will automatically be enforced to not generate mature content"
          },
          "popularityRank": {
            "type": "number",
            "description": "Get a rank between 0-1 on the popularity of the resource.",
            "format": "double",
            "nullable": true
          },
          "isFeatured": {
            "type": "boolean",
            "description": "Get wether this resource is featured",
            "nullable": true
          },
          "publishedAt": {
            "type": "string",
            "description": "The date at which this model got published",
            "format": "date-time",
            "nullable": true
          },
          "hasNSFWContentRestriction": {
            "type": "boolean",
            "description": "A boolean indicating whether this resource restricts to SFW content generation.\r\nNSFWContent covers X and AA whereas MatureContent includes R rated content."
          }
        },
        "additionalProperties": false,
        "description": "Details for a specific resource."
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "description": "Copy valid authorization token into field",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ],
...
}